feat: add rem,est wo column to table and refactor
